+++ package.json
@@ -25,6 +25,7 @@
     "react-scripts": "4.0.3",
     "redux": "^4.1.0",
     "redux-actions": "^2.6.5",
+    "redux-saga": "^1.1.3",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
